requirements.txt
numpy
Pillow
matplotlib
pycryptodome